+++ src/actions/stack_traversal.ts
